Fixed pylint errors separated TypeDict class in one file
